+++ src/test/java/io/supertokens/storage/mysql/test/DeadlockTest.java
@@ -24,6 +24,12 @@ import io.supertokens.pluginInterface.Storage;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.pluginInterface.sqlStorage.SQLStorage;
+import io.supertokens.pluginInterface.totp.TOTPDevice;
+import io.supertokens.pluginInterface.totp.TOTPUsedCode;
+import io.supertokens.pluginInterface.totp.exception.TotpNotEnabledException;
+import io.supertokens.pluginInterface.totp.exception.UsedCodeAlreadyExistsException;
+import io.supertokens.pluginInterface.totp.sqlStorage.TOTPSQLStorage;
+import io.supertokens.pluginInterface.sqlStorage.SQLStorage.TransactionIsolationLevel;
 import io.supertokens.storageLayer.StorageLayer;
 import org.junit.AfterClass;
 import org.junit.Before;
@@ -31,6 +37,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestRule;
 
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
@@ -40,6 +48,8 @@ import java.util.concurrent.atomic.AtomicReference;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+
+import static io.supertokens.storage.mysql.QueryExecutorTemplate.update;
 
 public class DeadlockTest {
     @Rule
@@ -223,6 +233,336 @@ public class DeadlockTest {
         assertNull(process
                 .checkOrWaitForEventInPlugin(io.supertokens.storage.mysql.ProcessState.PROCESS_STATE.DEADLOCK_FOUND));
         assert (pass.get());
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testConcurrentDeleteAndInsert() throws Exception {
+        String[] args = { "../" };
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        Storage storage = StorageLayer.getStorage(process.getProcess());
+        SQLStorage sqlStorage = (SQLStorage) storage;
+
+        // Create a device as well as a user:
+        TOTPSQLStorage totpStorage = StorageLayer.getTOTPStorage(process.getProcess());
+        TOTPDevice device = new TOTPDevice("user", "d1", "secret", 30, 1, false);
+        totpStorage.createDevice(device);
+
+        long now = System.currentTimeMillis();
+        long nextDay = now + 1000 * 60 * 60 * 24; // 1 day from now
+        TOTPUsedCode code = new TOTPUsedCode("user", "1234", true, nextDay, now);
+        totpStorage.startTransaction(con -> {
+            try {
+                totpStorage.insertUsedCode_Transaction(con, code);
+                totpStorage.commitTransaction(con);
+            } catch (TotpNotEnabledException | UsedCodeAlreadyExistsException e) {
+                // This should not happen
+                throw new StorageTransactionLogicException(e);
+            }
+            return null;
+        });
+
+        final Object syncObject = new Object();
+
+        AtomicReference<String> t1State = new AtomicReference<>("init");
+        AtomicReference<String> t2State = new AtomicReference<>("init");
+
+        AtomicBoolean t1Failed = new AtomicBoolean(true);
+        AtomicBoolean t2Failed = new AtomicBoolean(false);
+
+        Runnable r1 = () -> {
+            try {
+                sqlStorage.startTransaction(con -> {
+                    // Isolation level is SERIALIZABLE
+                    Connection sqlCon = (Connection) con.getConnection();
+
+                    String QUERY = "DELETE FROM totp_users where user_id = ?";
+                    try {
+                        update(sqlCon, QUERY, pst -> {
+                            pst.setString(1, "user");
+                        });
+                    } catch (SQLException e) {
+                        // Something is wrong with the test
+                        // This should not happen
+                        throw new StorageTransactionLogicException(e);
+                    }
+                    // Removal of user also triggers removal of the devices because
+                    // of FOREIGN KEY constraint.
+
+                    synchronized (syncObject) {
+                        // Notify t2 that that device has been deleted by t1
+                        t1State.set("query");
+                        syncObject.notifyAll();
+                    }
+
+                    // Wait for t2 to run the update the device query before committing
+                    synchronized (syncObject) {
+
+                        while (t2State.get().equals("init")) {
+                            try {
+                                syncObject.wait();
+                            } catch (InterruptedException ignored) {
+                            }
+                        }
+                    }
+
+                    sqlStorage.commitTransaction(con);
+                    t1State.set("commit");
+                    t1Failed.set(false);
+
+                    return null;
+                }, TransactionIsolationLevel.SERIALIZABLE);
+            } catch (StorageQueryException | StorageTransactionLogicException e) {
+                // This is expected because of "could not serialize access"
+                t1Failed.set(true);
+            }
+        };
+
+        Runnable r2 = () -> {
+            try {
+                sqlStorage.startTransaction(con -> {
+                    // Isolation level is SERIALIZABLE
+
+                    synchronized (syncObject) {
+                        // Wait for t1 to run delete device query first
+                        while (t1State.get().equals("init")) {
+                            try {
+                                syncObject.wait();
+                            } catch (InterruptedException ignored) {
+                            }
+                        }
+                    }
+
+                    Runnable r2Inner = () -> {
+                        // Wait for t2Inner to start running
+                        try {
+                            Thread.sleep(1500);
+                        } catch (InterruptedException ignored) {
+                        }
+
+                        // The psql txn will wait block t2Inner thread
+                        // but the t2 txn is still free and can commit.
+
+                        synchronized (syncObject) {
+                            // Notify t1 that that device has been updated by t2
+                            t2State.set("query");
+                            syncObject.notifyAll();
+                        }
+                    };
+
+                    Thread t2Inner = new Thread(r2Inner);
+                    t2Inner.start();
+                    // We will not wait for t2Inner to finish
+
+                    TOTPUsedCode code2 = new TOTPUsedCode("user", "1234", false, nextDay, now + 1);
+                    try {
+                        totpStorage.insertUsedCode_Transaction(con, code2);
+                    } catch (TotpNotEnabledException | UsedCodeAlreadyExistsException e) {
+                        // This should not happen
+                        throw new StorageTransactionLogicException(e);
+                    }
+                    sqlStorage.commitTransaction(con);
+                    t2State.set("commit");
+                    t2Failed.set(false);
+
+                    return null;
+                });
+            } catch (StorageTransactionLogicException e) {
+                Exception e2 = e.actualException;
+
+                if (e2 instanceof TotpNotEnabledException) {
+                    t2Failed.set(true);
+                }
+            } catch (StorageQueryException e) {
+                t2Failed.set(false);
+            }
+        };
+
+        Thread t1 = new Thread(r2);
+        Thread t2 = new Thread(r1);
+
+        t1.start();
+        t2.start();
+
+        t1.join();
+        t2.join();
+
+        // t1 (delete) should succeed
+        // but t2 (insert) should fail because of "could not serialize access"
+        assertTrue(!t1Failed.get() && t2Failed.get());
+        assert (t1State.get().equals("commit") && t2State.get().equals("query"));
+        assertNull(process
+                .checkOrWaitForEventInPlugin(io.supertokens.storage.mysql.ProcessState.PROCESS_STATE.DEADLOCK_FOUND,
+                        1000));
+
+        process.kill();
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));
+    }
+
+    @Test
+    public void testConcurrentDeleteAndUpdate() throws Exception {
+        String[] args = { "../" };
+        TestingProcessManager.TestingProcess process = TestingProcessManager.start(args);
+        assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STARTED));
+
+        Storage storage = StorageLayer.getStorage(process.getProcess());
+        SQLStorage sqlStorage = (SQLStorage) storage;
+
+        // Create a device as well as a user:
+        TOTPSQLStorage totpStorage = StorageLayer.getTOTPStorage(process.getProcess());
+        TOTPDevice device = new TOTPDevice("user", "d1", "secret", 30, 1, false);
+        totpStorage.createDevice(device);
+
+        long now = System.currentTimeMillis();
+        long nextDay = now + 1000 * 60 * 60 * 24; // 1 day from now
+        TOTPUsedCode code = new TOTPUsedCode("user", "1234", true, nextDay, now);
+        totpStorage.startTransaction(con -> {
+            try {
+                totpStorage.insertUsedCode_Transaction(con, code);
+                totpStorage.commitTransaction(con);
+            } catch (TotpNotEnabledException | UsedCodeAlreadyExistsException e) {
+                // This should not happen
+                throw new StorageTransactionLogicException(e);
+            }
+            return null;
+        });
+
+        final Object syncObject = new Object();
+
+        AtomicReference<String> t1State = new AtomicReference<>("init");
+        AtomicReference<String> t2State = new AtomicReference<>("init");
+
+        AtomicBoolean t1Failed = new AtomicBoolean(true);
+        AtomicBoolean t2Failed = new AtomicBoolean(false);
+
+        Runnable r1 = () -> {
+            try {
+                sqlStorage.startTransaction(con -> {
+                    // Isolation level is SERIALIZABLE
+                    Connection sqlCon = (Connection) con.getConnection();
+
+                    String QUERY = "DELETE FROM totp_users where user_id = ?";
+                    try {
+                        update(sqlCon, QUERY, pst -> {
+                            pst.setString(1, "user");
+                        });
+                    } catch (SQLException e) {
+                        // Something is wrong with the test
+                        // This should not happen
+                        throw new StorageTransactionLogicException(e);
+                    }
+                    // Removal of user also triggers removal of the devices because
+                    // of FOREIGN KEY constraint.
+
+                    synchronized (syncObject) {
+                        // Notify t2 that that device has been deleted by t1
+                        t1State.set("query");
+                        syncObject.notifyAll();
+                    }
+
+                    // Wait for t2 to run the update the device query before committing
+                    synchronized (syncObject) {
+
+                        while (t2State.get().equals("init")) {
+                            try {
+                                syncObject.wait();
+                            } catch (InterruptedException ignored) {
+                            }
+                        }
+                    }
+
+                    sqlStorage.commitTransaction(con);
+                    t1State.set("commit");
+                    t1Failed.set(false);
+
+                    return null;
+                }, TransactionIsolationLevel.SERIALIZABLE);
+            } catch (StorageQueryException | StorageTransactionLogicException e) {
+                // This is expected because of "could not serialize access"
+                t1Failed.set(true);
+            }
+        };
+
+        Runnable r2 = () -> {
+            try {
+                sqlStorage.startTransaction(con -> {
+                    // Isolation level is SERIALIZABLE
+                    Connection sqlCon = (Connection) con.getConnection();
+
+                    synchronized (syncObject) {
+                        // Wait for t1 to run delete device query first
+                        while (t1State.get().equals("init")) {
+                            try {
+                                syncObject.wait();
+                            } catch (InterruptedException ignored) {
+                            }
+                        }
+                    }
+
+                    Runnable r2Inner = () -> {
+                        // Wait for t2Inner to start running
+                        try {
+                            Thread.sleep(1500);
+                        } catch (InterruptedException ignored) {
+                        }
+
+                        // The psql txn will wait block t2Inner thread
+                        // but the t2 txn is still free and can commit.
+
+                        synchronized (syncObject) {
+                            // Notify t1 that that device has been updated by t2
+                            t2State.set("query");
+                            syncObject.notifyAll();
+                        }
+                    };
+
+                    Thread t2Inner = new Thread(r2Inner);
+                    t2Inner.start();
+                    // We will not wait for t2Inner to finish
+
+                    String QUERY = "UPDATE totp_used_codes SET is_valid=false WHERE user_id = ?";
+                    try {
+                        update(sqlCon, QUERY, pst -> {
+                            pst.setString(1, "user");
+                        });
+                    } catch (SQLException e) {
+                        throw new StorageTransactionLogicException(e);
+                    }
+
+                    sqlStorage.commitTransaction(con);
+                    t2State.set("commit");
+                    t2Failed.set(false);
+
+                    return null;
+                });
+            } catch (StorageTransactionLogicException e) {
+                Exception e2 = e.actualException;
+
+                if (e2 instanceof TotpNotEnabledException) {
+                    t2Failed.set(true);
+                }
+            } catch (StorageQueryException e) {
+                t2Failed.set(false);
+            }
+        };
+
+        Thread t1 = new Thread(r2);
+        Thread t2 = new Thread(r1);
+
+        t1.start();
+        t2.start();
+
+        t1.join();
+        t2.join();
+
+        // Both t1 (delete) and t2 (update) should pass
+        // Unlike Postgres, MySQL completes the delete. The update becomes ineffective.
+        assertTrue(!t1Failed.get() && !t2Failed.get());
+        assert (t1State.get().equals("commit") && t2State.get().equals("commit"));
 
         process.kill();
         assertNotNull(process.checkOrWaitForEvent(ProcessState.PROCESS_STATE.STOPPED));